back/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic[email]
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
aiofiles